+++ xtrack/beam_elements/slice_elements_thin.py
